first commit
This is my first commit in this repository!!

I didn't know that I would publish this project on Github, so I used the help of the AI that is integrated on VS code to help me and it has been quite good.

I've been kinda addicted to Pokémon TCG lately and I like seeing beautiful arts, that's why I started this project.

I hope I can succeed my wishes with this one.