huge update
Yeah, this is the biggest update so far. The AI managed to do what I was imagining. Of course I made some adjustments myself, but overall, the execution was made by the AI.

I started making some progress on the description of each card, but I'm face a problem that I need to fix soon: the space for each column and how to put a card below another and put a line to separate each other, like a grid.

I might need some AI help on that, and also, fix a lot of the card description later. I just wanted to commit a new version now.