Base Set cards completed!!
This is what I call as a update. I had some new ideas to give the page some identity coloring the types each Pokémon has weakness and resistance, and also their own type.

I also updated the image format to prioritize fast loading and the size of each image were changed to webp instead of png.

It was very fun to make this page and to see what card I find beautiful. Some choices had to be made haha