Fossil Set completed and some fixes!
I thought that I would only finish the set, but no! Everything had to be changed in some way.

I put the lightbox near the cards for each rarity, which saves me much time. I put the japanese pack as an image. For the first 3 packs, there were a special artwork for the 20th anniversary of the TCG and I put them at the bottom of the packs pages. Unfortunately there are not for the following packs.

I also put a favicon for the Jungle and Fossil sets. I need to think what image I will put for the base set.

I also fixed the description of what was naming Artist instead of Illustrator.

The release of the packs both in Japanese and in English were added at the presentation of each pack