Base 2 Set completed!!
I had some doubts about creating a page for the Base 2 set, but I saw a good opportunity to create something fun.

I put the Holo Rare cards that I had since the Base Set and Jungle Set and I also created a place where we can see the difference on each card. Not only the background was changed, but also some of the characteristcs of the cards were changed, which is cool to notice.

I also removed the animation I used of the transition for the normal cards. I was thinking that it would confuse more than to be an important feature.

There is a transition, but it is not transparent when you click to see the next/previous card. It's better to see than it was before.

I'll try to find a good transition to put in the future. I have an idea, but I couldn't find what I was looking for.

I thought that this set would be "the cards here appear in other sets. Come see them here" and link to the Base and Jungle sets, but I was completely wrong haha. It was fun though.